// backend-fashion/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const page_group = {
  about_us: 'about_us',
  resource: 'resource'
} as const

export type page_group = (typeof page_group)[keyof typeof page_group]


export const page_section_type = {
  hero: 'hero',
  content: 'content',
  feature: 'feature',
  gallery: 'gallery',
  faq: 'faq'
} as const

export type page_section_type = (typeof page_section_type)[keyof typeof page_section_type]


export const email_otps_type = {
  register: 'register',
  reset_password: 'reset_password'
} as const

export type email_otps_type = (typeof email_otps_type)[keyof typeof email_otps_type]


export const coupons_type = {
  percent: 'percent',
  fixed: 'fixed'
} as const

export type coupons_type = (typeof coupons_type)[keyof typeof coupons_type]


export const inventory_logs_type = {
  import: 'import',
  sale: 'sale',
  refund: 'refund',
  adjust: 'adjust'
} as const

export type inventory_logs_type = (typeof inventory_logs_type)[keyof typeof inventory_logs_type]


export const payments_method = {
  cod: 'cod',
  momo: 'momo',
  vnpay: 'vnpay',
  stripe: 'stripe'
} as const

export type payments_method = (typeof payments_method)[keyof typeof payments_method]


export const review_media_type = {
  image: 'image',
  video: 'video'
} as const

export type review_media_type = (typeof review_media_type)[keyof typeof review_media_type]


export const orders_status = {
  pending: 'pending',
  confirmed: 'confirmed',
  shipping: 'shipping',
  completed: 'completed',
  cancelled: 'cancelled'
} as const

export type orders_status = (typeof orders_status)[keyof typeof orders_status]


export const orders_payment_status = {
  unpaid: 'unpaid',
  paid: 'paid',
  failed: 'failed',
  refunded: 'refunded'
} as const

export type orders_payment_status = (typeof orders_payment_status)[keyof typeof orders_payment_status]


export const payments_status = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
  refunded: 'refunded'
} as const

export type payments_status = (typeof payments_status)[keyof typeof payments_status]


export const shipments_status = {
  preparing: 'preparing',
  shipped: 'shipped',
  delivered: 'delivered',
  returned: 'returned'
} as const

export type shipments_status = (typeof shipments_status)[keyof typeof shipments_status]


export const products_status = {
  active: 'active',
  inactive: 'inactive',
  out_of_stock: 'out_of_stock'
} as const

export type products_status = (typeof products_status)[keyof typeof products_status]